requirements.txt
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
biopython==1.81
bleach==6.1.0
boto3==1.28.62
botocore==1.31.62
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
brotlipy==0.7.0
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cloudpickle==3.0.0
cmake==3.27.0
comm==0.1.4
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
filelock @ file:///croot/filelock_1672387128942/work
flit_core @ file:///croot/flit-core_1679397103445/work/source/flit_core
fonttools==4.39.3
fqdn==1.5.1
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
huggingface-hub==0.13.4
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==6.8.0
importlib-resources==5.12.0
ipykernel==6.25.2
ipython==8.12.3
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1666908132255/work
jmespath==1.0.1
joblib @ file:///croot/joblib_1666298844297/work
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.4.0
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
Keras==2.0.6
kiwisolver==1.4.4
lightning-utilities==0.9.0
lit==16.0.6
llvmlite==0.41.1
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mpmath==1.2.1
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx @ file:///croot/networkx_1678964333703/work
notebook_shim==0.2.3
numba==0.58.1
numexpr @ file:///croot/numexpr_1668713893690/work
numpy @ file:///croot/numpy_and_numpy_base_1682520569166/work
openpyxl==3.0.10
overrides==7.4.0
packaging @ file:///croot/packaging_1678965309396/work
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
pooch @ file:///tmp/build/80754af9/pooch_1623324770023/work
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pygad==3.3.1
Pygments==2.16.1
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pyteomics==4.6
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-json-logger==2.0.7
pytorch-ranger @ file:///home/conda/feedstock_root/build_artifacts/pytorch-ranger_1618590827371/work
pytz @ file:///croot/pytz_1671697431263/work
PyYAML==6.0
pyzmq==25.1.1
referencing==0.30.2
regex==2023.3.23
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.4
s3transfer==0.7.0
scikit-learn @ file:///croot/scikit-learn_1680198742267/work
scipy==1.10.1
seaborn==0.13.0
Send2Trash==1.8.2
sentencepiece==0.1.99
shap==0.43.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
slicer==0.0.7
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy @ file:///croot/sympy_1668202399572/work
terminado==0.17.1
Theano==1.0.5
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1+cu118
torch-optimizer @ file:///home/conda/feedstock_root/build_artifacts/torch-optimizer_1635700648766/work
torchaudio==2.0.0
torchmetrics==1.2.0
torchvision==0.15.2+cu118
tornado==6.3.3
tqdm==4.65.0
traitlets==5.11.2
transformers==4.28.1
triton==2.0.0
types-python-dateutil==2.8.19.14
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1680254681959/work
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
zipp==3.15.0
